SCRUM-65 Restructure some components
