+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { UserContext } from "./UserContext";
 import Header from "./components/Header";
-import Garden from "./components/Garden";
+import Garden from "./components/gardenComponents/Garden";
 import Register from "./components/Register";
 import Logout from "./components/Logout";
 import Login from "./components/Login";
@@ -13,6 +13,7 @@ import Map from "./components/Map";
 />;
 
 import { User } from "./types/User";
+import HomePage from "./components/HomePage";
 
 function App() {
   const [user, setUser] = useState<User | null>(() => {
@@ -41,7 +42,8 @@ function App() {
         <div className="App">
           <Header title="Garden - nevem" />
           <Routes>
-            <Route path="/" element={<Map />} />
+            <Route path="/" element={<HomePage />} />
+            <Route path="/map" element={<Map />} />
             <Route path="/garden" element={<Garden />} />
             <Route path="/login" element={<Login />} />
             <Route path="/register" element={<Register />} />
