+++ frontend/src/components/Header.tsx
@@ -17,6 +17,9 @@ function Header({ title }: HeaderProps) {
           <li>
             <Link to="/">Home</Link>
           </li>
+          <li>
+            <Link to="/map">Map</Link>
+          </li>
           {user ? (
             <li>
               <Link to="/logout">Logout</Link>
@@ -32,12 +35,7 @@ function Header({ title }: HeaderProps) {
             </>
           )}
         </ul>
-        {user ?
-        <div>
-          Logged in User: {user.username}
-        </div>
-        : null}
-        
+        {user ? <div>Logged in User: {user.username}</div> : null}
       </nav>
     </header>
   );
