+++ frontend/src/components/HomePage.tsx
@@ -1,0 +1,9 @@
+function HomePage() {
+  return (
+    <div>
+      <h1>Welcome to home page!</h1>
+    </div>
+  );
+}
+
+export default HomePage;
